+++ ZEngine/ZEngine/Rendering/Renderers/Compute/FrustumCullingPass.cpp
@@ -1,0 +1,10 @@
+#include <ZEngine/Rendering/Renderers/Compute/FrustumCullingPass.h>
+
+namespace ZEngine::Rendering::Renderers
+{
+    void FrustumCullingPass::Setup(Hardwares::VulkanDevicePtr const device, cstring name, RenderGraphResourceBuilderPtr const res_builder, RenderGraphResourceInspectorPtr res_inspector) {}
+
+    void FrustumCullingPass::Compile(Hardwares::VulkanDevicePtr const device, Rendering::Scenes::SceneDataPtr const scene, RenderPasses::RenderPassBuilder* pass_builder, RenderGraphResourceInspectorPtr res_inspector, RenderPasses::RenderPass** const output_pass) {}
+
+    void FrustumCullingPass::Execute(Hardwares::VulkanDevicePtr const device, RenderGraphResourceInspectorPtr res_inspector, Rendering::Scenes::SceneDataPtr const scene, RenderPasses::RenderPass* const pass, Buffers::FramebufferVNext* const framebuffer, Hardwares::CommandBufferPtr const command_buffer) {}
+} // namespace ZEngine::Rendering::Renderers
