+++ ZEngine/ZEngine/Rendering/Renderers/Contracts/RendererDataContract.h
@@ -5,14 +5,10 @@ namespace ZEngine::Rendering::Renderers::Contracts
 {
     struct UBOCameraLayout
     {
-        alignas(16) ZEngine::Core::Maths::Mat4f View        = ZEngine::Core::Maths::Identity<ZEngine::Core::Maths::Mat4f>();
-        alignas(16) ZEngine::Core::Maths::Mat4f Projection  = ZEngine::Core::Maths::Identity<ZEngine::Core::Maths::Mat4f>();
-        alignas(16) ZEngine::Core::Maths::Vec4f Position    = ZEngine::Core::Maths::Vec4f(0.0f, 0.0f, 0.0f, 1.0f);
-        alignas(16) ZEngine::Core::Maths::Mat4f InvViewProj = ZEngine::Core::Maths::Identity<ZEngine::Core::Maths::Mat4f>();
+        ZEngine::Core::Maths::Mat4f View        = ZEngine::Core::Maths::Identity<ZEngine::Core::Maths::Mat4f>();
+        ZEngine::Core::Maths::Mat4f Projection  = ZEngine::Core::Maths::Identity<ZEngine::Core::Maths::Mat4f>();
+        ZEngine::Core::Maths::Vec4f Position    = ZEngine::Core::Maths::Vec4f(0.0f, 0.0f, 0.0f, 1.0f);
+        ZEngine::Core::Maths::Mat4f InvViewProj = ZEngine::Core::Maths::Identity<ZEngine::Core::Maths::Mat4f>();
     };
 
-    struct UBOModelLayout
-    {
-        alignas(16) ZEngine::Core::Maths::Mat4f Model = ZEngine::Core::Maths::Identity<ZEngine::Core::Maths::Mat4f>();
-    };
 } // namespace ZEngine::Rendering::Renderers::Contracts
