+++ ZEngine/ZEngine/Rendering/Renderers/Graphics/GbufferPass.h
@@ -1,0 +1,14 @@
+#pragma once
+#include <ZEngine/Rendering/Renderers/Base/RenderPass.h>
+#include <ZEngine/Rendering/Renderers/RenderGraph.h>
+#include <ZEngine/Rendering/Scenes/RenderScene.h>
+
+namespace ZEngine::Rendering::Renderers
+{
+    struct GbufferPass : public IRenderGraphCallbackPass
+    {
+        virtual void Setup(Hardwares::VulkanDevicePtr const device, cstring name, RenderGraphResourceBuilderPtr const res_builder, RenderGraphResourceInspectorPtr res_inspector) override;
+        virtual void Compile(Hardwares::VulkanDevicePtr const device, Rendering::Scenes::SceneDataPtr const scene, RenderPasses::RenderPassBuilder* pass_builder, RenderGraphResourceInspectorPtr res_inspector, RenderPasses::RenderPass** const output_pass) override;
+        virtual void Execute(Hardwares::VulkanDevicePtr const device, RenderGraphResourceInspectorPtr res_inspector, Rendering::Scenes::SceneDataPtr const scene, RenderPasses::RenderPass* const pass, Buffers::FramebufferVNext* const framebuffer, Hardwares::CommandBufferPtr const command_buffer) override;
+    };
+} // namespace ZEngine::Rendering::Renderers
