+++ ZEngine/ZEngine/Rendering/Renderers/Graphics/SkyboxPass.cpp
@@ -1,0 +1,115 @@
+#include <ZEngine/Engine.h>
+#include <ZEngine/Rendering/Renderers/Contracts/RendererDataContract.h>
+#include <ZEngine/Rendering/Renderers/GraphicRenderer.h>
+#include <ZEngine/Rendering/Renderers/Graphics/SkyboxPass.h>
+
+using namespace ZEngine::Helpers;
+using namespace ZEngine::Rendering::Specifications;
+using namespace ZEngine::Core::Containers;
+
+namespace ZEngine::Rendering::Renderers
+{
+    void SkyboxPass::Setup(Hardwares::VulkanDevicePtr const device, cstring name, RenderGraphResourceBuilderPtr const res_builder, RenderGraphResourceInspectorPtr res_inspector)
+    {
+        // DrawVertex layout: x y z nx ny nz u v (8 floats = 32 bytes)
+        // Normals and UVs zeroed — skybox shader only reads position (location 0, offset 0).
+        static constexpr float verts[] = {
+            -1.f, -1.f, 1.f, 0.f, 0.f, 0.f, 0.f, 0.f, 1.f, -1.f, 1.f, 0.f, 0.f, 0.f, 0.f, 0.f, 1.f, 1.f, 1.f, 0.f, 0.f, 0.f, 0.f, 0.f, -1.f, 1.f, 1.f, 0.f, 0.f, 0.f, 0.f, 0.f, -1.f, -1.f, -1.f, 0.f, 0.f, 0.f, 0.f, 0.f, 1.f, -1.f, -1.f, 0.f, 0.f, 0.f, 0.f, 0.f, 1.f, 1.f, -1.f, 0.f, 0.f, 0.f, 0.f, 0.f, -1.f, 1.f, -1.f, 0.f, 0.f, 0.f, 0.f, 0.f,
+        };
+        static constexpr uint32_t idxs[] = {0, 1, 2, 2, 3, 0, 1, 5, 6, 6, 2, 1, 5, 4, 7, 7, 6, 5, 4, 0, 3, 3, 7, 4, 3, 2, 6, 6, 7, 3, 4, 5, 1, 1, 0, 4};
+
+        auto*                     rrm    = ZEngine::Engine::GetContext()->RenderResourceManager;
+        ZENGINE_VALIDATE_ASSERT(rrm, "SkyboxPass::Setup: RenderResourceManager not available")
+        rrm->RegisterBuiltinGeometry(verts, sizeof(verts), idxs, 36, m_vtx_offset, m_idx_offset);
+
+        bool env_map_available = false;
+        if (EnvMapPath && EnvMapPath[0] != '\0')
+        {
+            auto* vfs = ZEngine::Engine::GetContext() ? ZEngine::Engine::GetContext()->VFS : nullptr;
+            if (vfs)
+            {
+                auto path_result   = ZEngine::Core::VFS::VFSPath::FromNative(EnvMapPath);
+                auto exists_result = path_result.Succeeded() ? vfs->Exists(path_result.Value()) : ZEngine::Core::VFS::VFSResult<bool>::Fail(ZEngine::Core::VFS::VFSError::InvalidPath);
+                if (exists_result.Failed() || !exists_result.Value())
+                    ZENGINE_CORE_ERROR("[SkyboxPass] Environment map not found in VFS: {}", EnvMapPath)
+                else
+                    env_map_available = true;
+            }
+            else
+            {
+                ZENGINE_CORE_ERROR("[SkyboxPass] VFS not available — cannot resolve environment map path: {}", EnvMapPath)
+            }
+        }
+
+        if (env_map_available)
+        {
+            auto* rrm = ZEngine::Engine::GetContext()->RenderResourceManager;
+            if (rrm)
+                m_env_map = rrm->SubmitTextureFile(EnvMapPath);
+        }
+
+        res_builder->ReadDepth(RendererResourceName::FrameDepthRenderTargetName);
+        res_builder->WriteColorAttachment(RendererResourceName::FrameColorRenderTargetName, {});
+    }
+
+    void SkyboxPass::Compile(Hardwares::VulkanDevicePtr const device, Rendering::Scenes::SceneDataPtr const scene, RenderPasses::RenderPassBuilder* pass_builder, RenderGraphResourceInspectorPtr res_inspector, RenderPasses::RenderPass** const output_pass)
+    {
+        CHECK_AND_ESCAPE_NULL(output_pass)
+
+        if (output_pass && !(*output_pass))
+        {
+            auto pass_spec = pass_builder->SetPipelineName("Skybox-Pipeline")
+                                 .SetInputBindingCount(1)
+                                 .SetStride(0, sizeof(float) * 8)
+                                 .SetRate(0, VK_VERTEX_INPUT_RATE_VERTEX)
+                                 .SetInputAttributeCount(1)
+                                 .SetLocation(0, 0)
+
+                                 .SetBinding(0, 0)
+                                 .SetFormat(0, Specifications::ImageFormat::R32G32B32_SFLOAT)
+                                 .SetOffset(0, 0)
+                                 .EnablePipelineDepthTest(true)
+                                 .EnablePipelineDepthWrite(false)
+                                 .PipelineDepthCompareOp(2)
+
+                                 .EnablePipelineBlending(false)
+
+                                 .SetCullMode(0)
+                                 .UseShader("skybox")
+                                 .Detach();
+            // clang-format off
+            *output_pass = device->CreateRenderPass(std::move(pass_spec));
+            // clang-format on
+            (*output_pass)->Bake();
+        }
+
+        if (scene && m_env_map.Valid())
+        {
+            (*output_pass)->SetDynamicUniform("UBCamera", sizeof(Contracts::UBOCameraLayout));
+            (*output_pass)->SetTexture("EnvMap", m_env_map);
+            (*output_pass)->SetSampler("LinearClampToEdgeSampler", device->GlobalLinearClampToEdgeSamplerImageInfo);
+            (*output_pass)->Verify();
+        }
+    }
+
+    void SkyboxPass::Execute(Hardwares::VulkanDevicePtr const device, RenderGraphResourceInspectorPtr res_inspector, Rendering::Scenes::SceneDataPtr const scene, RenderPasses::RenderPass* const pass, Buffers::FramebufferVNext* const framebuffer, Hardwares::CommandBufferPtr const command_buffer)
+    {
+        if (!m_env_map.Valid())
+            return;
+
+        command_buffer->BeginRenderPass(pass, framebuffer->Handle, false);
+        {
+            uint32_t w = pass->GetRenderAreaWidth();
+            uint32_t h = pass->GetRenderAreaHeight();
+            command_buffer->SetViewport(w, h);
+            command_buffer->SetScissor(w, h);
+        }
+        auto* rrm = ZEngine::Engine::GetContext()->RenderResourceManager;
+        command_buffer->BindPipeline(Specifications::PipelineBindPoint::GRAPHIC, pass->Pipeline);
+        command_buffer->BindVertexBuffer(*rrm->GetBuiltinVertexBuffer());
+        command_buffer->BindIndexBuffer(*rrm->GetBuiltinIndexBuffer(), VK_INDEX_TYPE_UINT32);
+        command_buffer->BindDescriptorSets(device->SwapchainPtr->CurrentFrame->Index, scene ? &scene->CameraHeapOffset : nullptr, scene ? 1u : 0u);
+        command_buffer->DrawIndexed(36, 1, m_idx_offset, static_cast<int32_t>(m_vtx_offset), 0);
+        command_buffer->EndRenderPass();
+    }
+} // namespace ZEngine::Rendering::Renderers
