+++ ZEngine/ZEngine/Rendering/Renderers/RenderGraph.h
@@ -4,7 +4,7 @@
 #include <ZEngine/Hardwares/DeferredFreeQueue.h>
 #include <ZEngine/Hardwares/VulkanDevice.h>
 #include <ZEngine/Rendering/Buffers/Framebuffer.h>
-#include <ZEngine/Rendering/Renderers/RenderPasses/RenderPass.h>
+#include <ZEngine/Rendering/Renderers/Base/RenderPass.h>
 #include <ZEngine/Rendering/Scenes/RenderScene.h>
 #include <ZEngine/Rendering/Specifications/TextureSpecification.h>
 #include <ZEngine/Rendering/Textures/Texture.h>
